backend/requirements.txt
fastapi
motor
python-dotenv
uvicorn
python-multipart
cloudinary
groq